requirements.txt
moviepy
opencv-python
torch 
torchvision
ultralytics
flask
ffmpeg-python